Convert the image into floating-point with 0-1 range.
Upon running the code provided as solution for the `Using a mask for a histogram (15 min)` section, the resulting histogram of the masked area doesn't look like what is expected. To reproduce the histogram, we need to convert the grayscale image of integer dtype, with 0-255 range, into a floating-point one with 0-1 range, by calling the function `skimage.util.img_as_float`.